fix: stub termion implementation on windows
