+++ examples/examples/termion.rs
@@ -1,40 +1,52 @@
-use std::io;
+#[cfg(not(windows))]
+mod example {
 
-use terminput::{Encoding, Event, KeyCode};
-use terminput_termion::to_terminput;
-use termion::input::{MouseTerminal, TermRead};
-use termion::raw::IntoRawMode;
+    use std::io;
 
-fn print_events() -> io::Result<()> {
-    let stdin = io::stdin();
-    let mut buf = [0; 16];
+    use terminput::{Encoding, Event, KeyCode};
+    use terminput_termion::to_terminput;
+    use termion::input::TermRead;
 
-    for event in stdin.events() {
-        let event: Result<Event, _> = to_terminput(event?);
+    pub fn print_events() -> io::Result<()> {
+        let stdin = io::stdin();
+        let mut buf = [0; 16];
 
-        if let Ok(event) = event {
-            println!("Event:   {:?}\r", event);
-            let written = event.encode(&mut buf, Encoding::Xterm);
-            if let Ok(written) = written {
-                println!("Encoded: {:?}\r", &buf[..written]);
-                if let Ok(Some(decoded)) = Event::parse_from(&buf[..written]) {
-                    println!("Decoded: {:?}\r", decoded);
+        for event in stdin.events() {
+            let event: Result<Event, _> = to_terminput(event?);
+
+            if let Ok(event) = event {
+                println!("Event:   {:?}\r", event);
+                let written = event.encode(&mut buf, Encoding::Xterm);
+                if let Ok(written) = written {
+                    println!("Encoded: {:?}\r", &buf[..written]);
+                    if let Ok(Some(decoded)) = Event::parse_from(&buf[..written]) {
+                        println!("Decoded: {:?}\r", decoded);
+                    }
+                }
+
+                if event == Event::Key(KeyCode::Esc.into()) {
+                    break;
                 }
             }
-
-            if event == Event::Key(KeyCode::Esc.into()) {
-                break;
-            }
+            println!();
         }
-        println!();
+        Ok(())
     }
+}
+
+#[cfg(not(windows))]
+fn main() -> std::io::Result<()> {
+    use termion::input::MouseTerminal;
+    use termion::raw::IntoRawMode;
+
+    let stdout = std::io::stdout().into_raw_mode()?;
+    let _stdout = MouseTerminal::from(stdout);
+    example::print_events()?;
+
     Ok(())
 }
 
-fn main() -> io::Result<()> {
-    let stdout = io::stdout().into_raw_mode()?;
-    let _stdout = MouseTerminal::from(stdout);
-    print_events()?;
-
-    Ok(())
+#[cfg(windows)]
+fn main() {
+    panic!("termion is not supported on windows");
 }
